--- SurveyQuestionFactory.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace Database\Factories;
-
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Model>
- */
-class SurveyQuestionFactory extends Factory
-{
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition(): array
-    {
-        return [
-            'question' => $this->faker->sentence,
-            'left_side_text' => $this->faker->word,
-            'right_side_text' => $this->faker->word,
-        ];
-    }
-}
