--- SurveySurveyQuestionFactory.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace Database\Factories;
-
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Model>
- */
-class SurveySurveyQuestionFactory extends Factory
-{
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition(): array
-    {
-        return [
-            'questionnaire_id' => Questionnaire::factory(),
-            'survey_question_id' => SurveyQuestion::factory(),
-        ];
-    }
-}
