+++ database/factories/SurveyAnswerFactory.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Database\Factories;
+
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Model>
+ */
+class SurveyAnswerFactory extends Factory
+{
+    /**
+     * Define the model's default state.
+     *
+     * @return array<string, mixed>
+     */
+    public function definition(): array
+    {
+        return [
+            'chosen_answer' => $this->faker->numberBetween(1, 5),
+            'survey_id' => $this->faker->numberBetween(1, 30),
+            'survey_question_id' => $this->faker->numberBetween(1, 30),
+        ];
+    }
+}
