+++ database/seeders/DatabaseSeeder.php
@@ -2,8 +2,15 @@
 
 namespace Database\Seeders;
 
+use App\Models\Questionnaire;
 use App\Models\User;
+use App\Models\Role;
+use App\Models\SurveySurveyQuestion;
+use App\Models\SurveyAnswer;
+use App\Models\Survey;
 // use Illuminate\Database\Console\Seeds\WithoutModelEvents;
+use Database\Factories\SurveyQuestionFactory;
+use Database\Factories\SurveySurveyQuestionFactory;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -14,10 +21,10 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         // User::factory(10)->create();
-
-        User::factory()->create([
-            'name' => 'Test User',
-            'email' => 'test@example.com',
-        ]);
+        Role::factory()->create();
+        User::factory(30)->create();
+        Questionnaire::factory(30)->create();
+        SurveySurveyQuestion::factory(30)->create();
+        SurveyAnswer::factory(500)->create();
     }
 }
